change dateTime field to time.Time

+++ db/models/electric_field_model.go
@@ -9,7 +9,7 @@ import (
 type ElectricField struct {
 	gorm.Model
 	DateTime      time.Time `gorm:"<-:create;not null"`
-	ElectricField float32   `gorm:"<-:create;not null"`
+	ElectricField float64   `gorm:"<-:create;not null"`
 	RotorStatus   bool      `gorm:"<-:create;not null"`
 	PlaceID       int
 	Place         Place `gorm:"constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
